feat: Split the code in different files
