+++ requirements.txt
@@ -1,4 +1,3 @@
 ultralytics
 opencv-python-headless
 customtkinter
-pyinstaller
